Simple retry for DELETE requests
Right now the API does not raises 404 errors when we delete a resource
that does not exists. Thus we can retry delete requests until we get a
200.

A correct implementation would be to catch 404 errors as "delete
success", which is not needed now and complexify the retry
implementation (special retry case only for delete requests).

A test `test_delete_on_hubstorage_api_does_not_404` will be triggered
when the assumption does not hold anymore and the retry policy should be
fixed.

Signed-off-by: lsenta <laurent.senta@gmail.com>