Move idempotence dispatch to client and GET default to apiget
Change where the dispatch occurs so that _iter_lines simply forward
the request configuration to the client. The client decides whether
to apply the retry policy depending on request args.

Move the default GET behavior to apiget so that iter_json method is not
altered by this change. (it already implements a retry that may be
incompatible).

Signed-off-by: lsenta <laurent.senta@gmail.com>